allow to measure only one node
This commit allows the user to measure only one node instead of running
the crawler and measure every node found. The user can do so by
specifying the "-enr" flag.

In order to run the crawler, now the user has to specify the "-crawl"
flag. Otherwise, the program will output an error.